ui: screens: move into separate subfolder
This makes it easier to tell apart the different screens from e.g.
the widgets or framebuffer code that's located in ui as well.
